Add -Wshorten-64-to-32 and -Wtautological-unsigned-zero-compare to `inside_blink` config
Bug: 514751964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8062845
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1659987}

+++ net/quic/quic_proxy_datagram_client_socket.h
@@ -181,7 +181,7 @@ class NET_EXPORT_PRIVATE QuicProxyDatagramClientSocket
   // this by simply using the current chain and indexing the last proxy in
   // that chain.
   const ProxyChain& proxy_chain() { return proxy_chain_; }
-  int proxy_chain_index() { return proxy_chain_.length() - 1; }
+  size_t proxy_chain_index() { return proxy_chain_.length() - 1; }
 
   State next_state_ = STATE_DISCONNECTED;
